DEV-21 Changed folder structure
