+++ src/Shared/ExcelAdapter/domain/aliases.go
@@ -1,0 +1,4 @@
+package domain
+
+type Name = string
+type Id = uint16
