+++ src/Shared/ExcelAdapter/domain/structure.go
@@ -1,0 +1,13 @@
+package domain
+
+type Structure struct {
+	PacketName   string
+	Measurements []string
+}
+
+func newStructure(column []string) Structure {
+	return Structure{
+		PacketName:   column[0],
+		Measurements: column[1:],
+	}
+}
