--- src/Shared/ExcelParser/domain/board/aliases.go
@@ -1,4 +1,0 @@
-package board
-
-type Name = string
-type Id = uint16
