--- src/Shared/ExcelParser/domain/document/document.go
@@ -1,5 +1,0 @@
-package document
-
-type Document struct {
-	Sheets map[string]Sheet
-}
