--- src/Shared/ExcelParser/domain/document/row.go
@@ -1,5 +1,0 @@
-package document
-
-type Row = []Cell
-
-type Cell = string
