--- src/Shared/ExcelParser/domain/document/sheet.go
@@ -1,6 +1,0 @@
-package document
-
-type Sheet struct {
-	Name   string
-	Tables map[string]Table
-}
