--- src/Shared/ExcelParser/domain/document/table.go
@@ -1,6 +1,0 @@
-package document
-
-type Table struct {
-	Name string
-	Rows []Row
-}
