+++ src/Shared/ExcelRetriever/domain/document.go
@@ -1,0 +1,5 @@
+package domain
+
+type Document struct {
+	Sheets map[string]Sheet
+}
