+++ src/Shared/ExcelRetriever/domain/row.go
@@ -1,0 +1,5 @@
+package domain
+
+type Row = []Cell
+
+type Cell = string
