+++ src/Shared/ExcelRetriever/domain/sheet.go
@@ -1,0 +1,6 @@
+package domain
+
+type Sheet struct {
+	Name   string
+	Tables map[string]Table
+}
