+++ src/Shared/ExcelRetriever/domain/table.go
@@ -1,0 +1,6 @@
+package domain
+
+type Table struct {
+	Name string
+	Rows []Row
+}
