+++ src/Shared/PacketAdapter/PacketParser/domain/enum.go
