+++ src/Shared/PacketAdapter/PacketParser/domain/measurement_data.go
@@ -1,0 +1,13 @@
+package domain
+
+type MeasurementData struct {
+	Name      string
+	ValueType string
+}
+
+func NewMeasurement(name string, valueType string) MeasurementData {
+	return MeasurementData{
+		Name:      name,
+		ValueType: valueType,
+	}
+}
