+++ src/Shared/PacketAdapter/TransportController/tests/resources/mixed.txt
